requirements.txt
tk
customtkinter
pdf2image
pillow
pymongo